Toggable light/dark style from menu
Included an alternate dark style based on Arc Dark theme and a new
menu item to toggle the style from light to dark and back.

The images have been changed to have transparency, so they look good in
both modes. A new image for the icon preserving the white background is
included.