Loads all the sound and startup and rewinds sounds
This fixes a bug in the ASFML version that causes that once a sound is
played, it doesn't play again (we have to rewind it).

There are changes also to load everything at startup. In fact, there was
no memory leak. When all the sound files are loaded, memory stops
increasing. This still happens, probably because the OS does not actually
load all in memory until it is used.

Closes #4